feat(suite-native): eject single wallet
